Updated the DTS bootargs to make the first block writeable
We don’t actually use the bootargs string in the DTS, but NBoot isn’t committed anywhere and this is a reasonable place for keeping the official copy.

The first block of flash has been made writeable as a hack for resetting the part into bootloader mode. If the first block doesn’t contain valid instructions, the bootloader will stay in loader mode instead of loading the firmware.